Finished Number Guessing Game
Added a while loop. Imported random. We type out a max number. Then we start guessing what the random number is between 0 and the max number the user specified. If it matches then the game will end.

I used if, elif and else statements. I also did a check for a Value Error